Allow CHTC pilots to utilize the syslog service.
This examines the JWT and, if it's a token from CHTC, it will steer
the authentication check to the CHTC collector.  This allows CHTC
pilots to utilize the OSG service.

This is mostly done to permit the CHTC pool to be used as a testing
ground for syslog integration before pushing changes to the OSPool.