Refer to the new primary OSPool CM (SOFTWARE-5005)
We actually have two CMs but it's not clear how to handle fallback in
the case of extended cm-1 failure but not cause issues when there are
intermittent cm-1 failures. E.g. if cm-1 is down when making the token
request and we fall back to cm-2 but cm-1 is up for retrieving the
approved token, I suspect that the end-to-end process would fail